planner: fix nulleq condition to calcalate index range
